+++ src/components/form/field/components/autocomplete.js
@@ -15,7 +15,7 @@ var _autocomplete = require("../../../form/autocomplete");
 
 var FieldAutocomplete = function FieldAutocomplete(props) {
   return _react.default.createElement(_marger.Marger, {
-    top: "1.5"
+    top: "1"
   }, _react.default.createElement(_autocomplete.Autocomplete, props));
 };