+++ src/components/form/field/components/password.js
@@ -15,7 +15,7 @@ var _passwordInput = require("../../password-input");
 
 var FieldPassword = function FieldPassword(props) {
   return _react.default.createElement(_marger.Marger, {
-    top: "1.5"
+    top: "1"
   }, _react.default.createElement(_passwordInput.PasswordInput, props));
 };